+++ storybook/src/components/document-card/document-card.content.ts
@@ -1,4 +1,4 @@
-import { InfoButton } from "dso-toolkit";
+import { InfoButton, Label } from "dso-toolkit";
 import { TemplateResult, html } from "lit-html";
 
 import { Templates } from "../../templates";
@@ -8,10 +8,18 @@ export const typeItems = [
   html`<span class="dso-document-card-type-item">Gemeente Amsterdam</span>`,
 ];
 
-const labels = ["BouwWerken", "Kleine windturbine"];
-export const labelsItems = labels.map(
-  (label) => html` <span style="border: 1px solid #333; border-radius: 4px; padding: 0 8px;">${label}</span> `,
-);
+export const labels: Label[] = [
+  {
+    label: "Bouwwerken",
+    compact: true,
+    status: "bright",
+  },
+  {
+    label: "Kleine windturbine",
+    compact: true,
+    status: "bright",
+  },
+];
 
 export function infoButton({ richContentTemplate }: Templates): InfoButton<TemplateResult> {
   return {